requirements-dev.txt
# Development and Testing Dependencies
-r requirements.txt

# Testing
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
factory-boy>=3.2.0

# Code Quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Development Tools
django-debug-toolbar>=4.0.0
django-extensions>=3.2.0

# Performance Monitoring
django-silk>=5.0.0

# Environment Management
python-dotenv>=1.0.0 